+++ url2parquet/url2parquet_download.py
@@ -9,7 +9,7 @@ from datetime import datetime, timedelta
 from tqdm import tqdm
 from glob import glob
 from io import BytesIO
-from nc_reader_download import nc_reader
+from url2parquet.nc_read_downloaded import nc_reader
 
 
 # 이미지 데이터를 Parquet로 저장하는 함수
@@ -122,7 +122,6 @@ def main():
     start_time = time.time()  # 시작 시간 기록
 
     data_types = ["VI004", "VI005", "VI006", "VI008", "NR013", "NR016", "SW038", "WV063", "WV069", "WV073", "IR087", "IR096", "IR105", "IR112", "IR123", "IR133"]
-    region = "KO"
     start_date = datetime(2025, 1, 10)  # 검색 시작일
     end_date = datetime.now().replace(hour=0, minute=0, second=0, microsecond=0)
     parquet_directory = "D:/sat_file/temp_parquet_hourly"  # 임시 시간별 parquet파일 저장 경로
